+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/logic/channel/internal/ChannelLogic.kt
@@ -153,7 +153,8 @@ internal class ChannelLogic(
             configs = configs,
             users = users.values.toList(),
             channels = listOf(channel),
-            messages = channel.messages
+            messages = channel.messages,
+            cacheForMessages = true
         )
     }