+++ stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/utils/extensions/Message.kt
@@ -19,6 +19,7 @@ package com.getstream.sdk.chat.utils.extensions
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.MessageSyncType
+import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import java.util.Date
@@ -32,21 +33,11 @@ internal fun Message.isDeleted(): Boolean = deletedAt != null
 /**
  * @return if the message was sent by current user.
  */
-@Deprecated(
-    "use isMine version with ChatClient parameter",
-    ReplaceWith(
-        "Message.isMine(chatClient: ChatClient)",
-        "io.getstream.chat.android.client.ChatClient"
-    )
-)
-@InternalStreamChatApi
-public fun Message.isMine(): Boolean = ChatClient.instance().getCurrentUser()?.id == user.id
-
-/**
- * @return if the message was sent by current user.
- */
 @InternalStreamChatApi
 public fun Message.isMine(chatClient: ChatClient): Boolean = chatClient.getCurrentUser()?.id == user.id
+
+@InternalStreamChatApi
+public fun Message.isMine(currentUser: User?): Boolean = currentUser?.id == user.id
 
 /**
  * @return when the message was created or throw an exception.
@@ -67,19 +58,19 @@ public fun Message.getCreatedAtOrNull(): Date? {
  * @return if the message failed at moderation or not.
  */
 @Deprecated(
-    "use isModerationFailed version with ChatClient parameter",
+    "use isModerationFailed version with `currentUser: User?` parameter",
     ReplaceWith(
-        "Message.isModerationFailed(chatClient: ChatClient)",
-        "io.getstream.chat.android.client.ChatClient"
+        "message.isModerationFailed(currentUser = chatClient.getCurrentUser())",
+        "io.getstream.chat.android.client.models.User"
     )
 )
-public fun Message.isModerationFailed(): Boolean = isMine() &&
+public fun Message.isModerationFailed(chatClient: ChatClient): Boolean = isMine(chatClient) &&
     syncStatus == SyncStatus.FAILED_PERMANENTLY &&
     syncDescription?.type == MessageSyncType.FAILED_MODERATION
 
 /**
  * @return if the message failed at moderation or not.
  */
-public fun Message.isModerationFailed(chatClient: ChatClient): Boolean = isMine(chatClient) &&
+public fun Message.isModerationFailed(currentUser: User?): Boolean = isMine(currentUser) &&
     syncStatus == SyncStatus.FAILED_PERMANENTLY &&
     syncDescription?.type == MessageSyncType.FAILED_MODERATION